requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.1
flake8==7.0.0
flasgger==0.9.7.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==3.0.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
langchain==0.0.315
langsmith==0.0.92
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
MarkupSafe==2.0.1
marshmallow==3.21.3
mccabe==0.7.0
mistune==3.0.2
more-itertools==8.10.0
multidict==6.0.5
mypy==0.942
mypy-extensions==0.4.3
mysql-connector-python==8.4.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openai==0.28.1
packaging==23.2
pathlib2==2.3.7.post1
pep8==1.7.1
psutil==5.9.0
pycodestyle==2.6.0
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.1
redis==5.0.4
referencing==0.33.0
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.0
SecretStorage==3.3.1
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
spotipy==2.23.0
SQLAlchemy==2.0.27
starkbank-ecdsa==2.2.0
systemd-python==234
tenacity==8.4.2
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.4
typed-ast==1.4.3
types-aiofiles==0.1
types-annoy==1.17
types-appdirs==1.4
types-atomicwrites==1.4
types-aws-xray-sdk==2.8
types-babel==2.9
types-backports-abc==0.5
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.10
types-bleach==4.1
types-boto==2.49
types-braintree==4.11
types-cachetools==4.2
types-caldav==0.8
types-certifi==2020.4
types-characteristic==14.3
types-chardet==4.0
types-click==7.1
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-contextvars==0.1
types-croniter==1.0
types-cryptography==3.3
types-dataclasses==0.1
types-dateparser==1.0
types-DateTimeRange==0.1
types-decorator==0.1
types-Deprecated==1.2
types-docopt==0.6
types-docutils==0.17
types-editdistance==0.5
types-emoji==1.2
types-entrypoints==0.3
types-enum34==1.1
types-filelock==3.2
types-first==2.0
types-Flask==1.1
types-freezegun==1.1
types-frozendict==0.1
types-futures==3.3
types-html5lib==1.1
types-httplib2==0.19
types-humanfriendly==9.2
types-ipaddress==1.0
types-itsdangerous==1.1
types-JACK-Client==0.1
types-Jinja2==2.11
types-jmespath==0.10
types-jsonschema==3.2
types-Markdown==3.3
types-MarkupSafe==1.1
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.0
types-oauthlib==3.1
types-orjson==3.6
types-paramiko==2.7
types-Pillow==8.3
types-polib==1.1
types-prettytable==2.1
types-protobuf==3.17
types-psutil==5.8
types-psycopg2==2.9
types-pyaudio==0.2
types-pycurl==0.1
types-pyfarmhash==0.2
types-Pygments==2.9
types-PyMySQL==1.0
types-pyOpenSSL==20.0
types-pyRFC3339==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-nmap==0.6
types-python-slugify==5.0
types-pytz==2021.1
types-pyvmomi==7.0
types-PyYAML==5.4
types-redis==3.5
types-requests==2.25
types-retry==0.9
types-selenium==3.141
types-Send2Trash==1.8
types-setuptools==57.4
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-stripe==2.59
types-tabulate==0.8
types-termcolor==1.1
types-toml==0.10
types-toposort==1.6
types-ttkthemes==3.2
types-typed-ast==1.4
types-tzlocal==0.1
types-ujson==0.1
types-vobject==0.9
types-waitress==0.1
types-Werkzeug==1.0
types-xxhash==2.0
typing-inspect==0.9.0
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.9.4
zipp==1.0.0
